requirements.txt
flask
python-dotenv
pydantic
numpy
pandas
scikit-learn
markdown
pyyaml
pytest
pyttsx3
requests
